pyproject.toml
[project]
name = "health-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.3.4",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.122.0",
    "google-generativeai>=0.8.5",
    "langchain>=1.0.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-groq>=1.0.0",
    "mysql-connector-python>=9.5.0",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pymupdf>=1.26.6",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "snowflake-connector-python>=4.1.0",
    "streamlit>=1.51.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.38.0",
]
